+++ CHANGELOG.md
@@ -2,8 +2,8 @@ Cinecred Changelog
 ==================
 
 
-1.7.0-SNAPSHOT
---------------
+1.7.0
+-----
 
 ### Additions